moved agitator code to new package
