+++ src/main/java/frc/robot/OI.java
@@ -14,7 +14,8 @@ import edu.wpi.first.wpilibj2.command.button.JoystickButton;
 import frc.robot.commands.control_panel.*;
 import frc.robot.commands.hopper.*;
 import frc.robot.commands.shooter.*;
-import frc.robot.commands.*;
+import frc.robot.commands.agitator.AgitatorRun;
+import frc.robot.commands.agitator.AgitatorStop;
 import frc.robot.commands.auto_commands.AlignRobot;
 
 /**
