+++ src/main/java/frc/robot/commands/agitator/AgitatorRun.java
@@ -1,4 +1,4 @@
-package frc.robot.commands;
+package frc.robot.commands.agitator;
 
 import frc.robot.Robot;
 import edu.wpi.first.wpilibj2.command.CommandBase;
